+++ packages/react/src/index.ts
@@ -16,6 +16,7 @@ export * from "./radialMenu";
 export * from "./quickMenu";
 export * from "./questTracker";
 export * from "./notifications";
+export * from "./fastTravel";
 export * from "./components";
 export * from "./hud";
 export * from "./bars";